Reorganize a bunch of shit
