+++ site/jsconfig.json
@@ -5,7 +5,6 @@
       "@assets/*": ["./assets/*"],
       "@components/*": ["./components/*"],
       "@hooks/*": ["./hooks/*"],
-      "@layout/*": ["./layout/*"],
       "@style/*": ["./style/*"],
       "@utils/*": ["./utils/*"]
     }
