+++ site/package.json
@@ -19,6 +19,7 @@
     "@mdx-js/react": "^1.6.22",
     "babel-plugin-styled-components": "^1.13.3",
     "date-fns": "^2.25.0",
+    "fuse.js": "^6.4.6",
     "gatsby": "^4.0.1",
     "gatsby-plugin-alias-imports": "^1.0.5",
     "gatsby-plugin-google-analytics": "^4.0.0",
@@ -34,6 +35,7 @@
     "gatsby-source-filesystem": "^4.0.0",
     "gatsby-transformer-sharp": "^4.0.0",
     "lodash": "^4.17.21",
+    "match-sorter": "^6.3.1",
     "parse-numeric-range": "^1.3.0",
     "prism-react-renderer": "^1.2.1",
     "react": "^17.0.1",
