+++ site/src/components/InlineCode/InlineCode.style.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const InlineCode = styled.code`
+  background-color: var(--bg-inline-code);
+  padding: 0.25rem;
+  word-break: break-word;
+`;
