+++ site/src/components/Link/Link.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import * as S from './style';
+import * as S from './Link.style';
 
 export const Link = (props) => {
   return props.to ? <S.InternalLink {...props} /> : <S.ExternalLink {...props} />;
