+++ site/src/components/Link/index.js
@@ -1,0 +1,1 @@
+export * from './Link';
