+++ site/src/components/Logo/Logo.js
@@ -1,40 +1,5 @@
-import styled from 'styled-components';
+import React from 'react';
 
-import { breakpoint, transition } from '@utils/mixins';
+import * as S from './Logo.style';
 
-import LogoSVG from '@assets/svg/logo.svg';
-
-const getWidth =
-  (bp) =>
-  ({ width }) => {
-    switch (bp) {
-      case breakpoint.md:
-        return `${width + width * 0.5}rem`;
-      case breakpoint.lg:
-        return `${width + width * 1.0}rem`;
-      default:
-        return `${width}rem`;
-    }
-  };
-
-export const Logo = styled(LogoSVG)`
-  width: ${getWidth()};
-  height: ${getWidth()};
-  cursor: pointer;
-  transition: ${transition('fill', 'stroke')};
-
-  g#logoFillStroke {
-    fill: var(--logo-fill);
-    stroke: var(--logo-stroke);
-  }
-
-  ${breakpoint.md} {
-    width: ${getWidth(breakpoint.md)};
-    height: ${getWidth(breakpoint.md)};
-  }
-
-  ${breakpoint.lg} {
-    width: ${getWidth(breakpoint.lg)};
-    height: ${getWidth(breakpoint.lg)};
-  }
-`;
+export const Logo = (props) => <S.Logo {...props} />;
