+++ site/src/components/Logo/index.js
@@ -1,0 +1,1 @@
+export * from './Logo';
