+++ site/src/components/Search/index.js
@@ -1,0 +1,1 @@
+export * from './Search';
