+++ site/src/components/SocialIcon/SocialIcon.style.js
