+++ site/src/components/Tag/Tag.js
@@ -1,13 +1,5 @@
-import styled from 'styled-components';
+import React from 'react';
 
-export const StyledTag = styled.div`
-  font-size: 0.75rem;
-  text-transform: uppercase;
-  border: solid 1.5px var(--primary);
-  font-weight: 700;
-  margin: 0 0.25rem 0.25rem 0;
-  padding: 0.5rem;
-  white-space: nowrap;
-`;
+import * as S from './Tag.style';
 
-export const Tag = StyledTag;
+export const Tag = (props) => <S.Tag {...props} />;
