+++ site/src/components/Tag/Tag.style.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+export const Tag = styled.div`
+  font-size: 0.75rem;
+  text-transform: uppercase;
+  border: solid 1.5px var(--primary);
+  font-weight: 700;
+  margin: 0 0.25rem 0.25rem 0;
+  padding: 0.5rem;
+  white-space: nowrap;
+`;
