--- site/src/components/index.js
@@ -1,12 +1,0 @@
-// This is more than just a barrell.  These components will all be imported to MDX and usable
-// in any MDX documents I write
-
-export * from './Button/Button';
-export * from './Callout/Callout';
-export * from './Card/Card';
-export * from './Code/Code';
-export * from './InlineCode/InlineCode';
-export * from './Link/Link';
-export * from './Logo/Logo';
-export * from './SocialIcon/SocialIcon';
-export * from './Tag/Tag';
