+++ site/src/pages/404.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 import styled from 'styled-components';
 
-import { Layout } from '@layout';
+import { Layout } from '@components/Layout';
 
 import FourOhFourSVG from '@assets/svg/hero/fourohfour.svg';
