+++ site/src/pages/_test.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function TestPage() {
+  return <div>Testing 123</div>;
+}
