+++ site/src/pages/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 import styled from 'styled-components';
 
-import { Layout } from '@layout';
+import { Layout } from '@components/Layout';
 import { transition } from '@utils/mixins';
 
 import DeskSVG from '@assets/svg/hero/desk.svg';
